+++ publisher/src/router/Router.tsx
@@ -19,8 +19,8 @@ import React from "react";
 import { Navigate, Route, Routes, useParams } from "react-router-dom";
 
 import { DataUpload } from "../components/DataUpload";
-import SpreadsheetReview from "../components/DataUpload/SpreadsheetReview";
-import UploadReview from "../components/DataUpload/UploadReview";
+import ShareSpreadsheetReview from "../components/DataUpload/ShareSpreadsheet";
+import UploadSpreadsheetReview from "../components/DataUpload/UploadSpreadsheet";
 import { MetricsDataChart } from "../components/DataViz/MetricsDataChart";
 import { REPORTS_LOWERCASE } from "../components/Global/constants";
 import { Guidance } from "../components/Guidance";
@@ -91,10 +91,13 @@ export const Router = () => {
           <Route path="/metric-config" element={<MetricsConfiguration />} />
           <Route path="/settings/*" element={<Settings />} />
           <Route path="/upload" element={<DataUpload />} />
-          <Route path="/upload/review-metrics" element={<UploadReview />} />
+          <Route
+            path="/upload/review-metrics"
+            element={<UploadSpreadsheetReview />}
+          />
           <Route
             path="/upload/:spreadsheetId/review-metrics"
-            element={<SpreadsheetReview />}
+            element={<ShareSpreadsheetReview />}
           />
 
           <Route
